+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Don't discard actor if it has more than 10 aliases.
 - Return instance actor JRD if Webfinger is queried with instance base URL.
 - Validate port number when parsing 'http' URLs.
+- Replace "same authority" checks with "same origin".
 
 ## [2.25.1] - 2024-07-11